Feature/sx1278 (#1)
Added sx1278 lora files from prior repo

Added the sx1278 unit test files and also  the testing scrip I used for that project becasue much of the hardware will be similar so its a good boiler plate

+++ src/devices/LoRa/sx1278/mySX1278.c
@@ -102,7 +102,7 @@ int activate_lora(int spi_handle)
     {
         return n;
     }
-    n = poll_reg(spi_handle, REG_OPMODE, 0xFF, mode, 100, 1000);
+    n = poll_reg(spi_handle, REG_OPMODE, (uint8_t)0xFF, mode, 100, 1000);
     return n;
 }
 
@@ -133,7 +133,7 @@ int deactivate_lora(int spi_handle)
     {
         return n;
     }
-    n = poll_reg(spi_handle, REG_OPMODE, 0xFF, mode, 100, 1000);
+    n = poll_reg(spi_handle, REG_OPMODE, (uint8_t)0xFF, mode, 100, 1000);
     return n;
 }
 
@@ -208,7 +208,7 @@ int set_sleep_mode(int spi_handle)
     {
         return n;
     }
-    n = poll_reg(spi_handle, REG_OPMODE, 0xFF, mode, 100, 1000);
+    n = poll_reg(spi_handle, REG_OPMODE, (uint8_t)0xFF, mode, 100, 1000);
     return n;
 }
 
@@ -226,7 +226,7 @@ int set_stdby_mode(int spi_handle)
     {
         return n;
     }
-    n = poll_reg(spi_handle, REG_OPMODE, 0xFF, mode, 100, 1000);
+    n = poll_reg(spi_handle, REG_OPMODE, (uint8_t)0xFF, mode, 100, 1000);
     return n;
 }
